Add files via upload
This report is based on the assignment after completion of course named "Microsoft Power BI for Business Intelligence" from Maven Analytics available on Udemy. The dataset was given by them only.
This report includes the use of slicer and the interactions gets changed according to need, KPI's and many more stuff's are available.